Split additional impls for Linear into a separate module
