+++ rust/ommx/src/convert/linear.rs
@@ -1,0 +1,30 @@
+use crate::v1::{linear::Term, Linear};
+use std::collections::BTreeSet;
+
+impl Linear {
+    pub fn new(terms: impl Iterator<Item = (u64, f64)>, constant: f64) -> Self {
+        Self {
+            terms: terms
+                .map(|(id, coefficient)| Term { id, coefficient })
+                .collect(),
+            constant,
+        }
+    }
+
+    pub fn single_term(id: u64, coefficient: f64) -> Self {
+        Self {
+            terms: vec![Term { id, coefficient }],
+            constant: 0.0,
+        }
+    }
+
+    pub fn used_decision_variable_ids(&self) -> BTreeSet<u64> {
+        self.terms.iter().map(|term| term.id).collect()
+    }
+}
+
+impl From<u64> for Linear {
+    fn from(id: u64) -> Self {
+        Self::single_term(id, 1.0)
+    }
+}
